Issue #37: Implement confirm-email endpoint
