+++ api/src/routes/app.rs
@@ -1,7 +1,11 @@
 use axum::Router;
 use sqlx::{Pool, Postgres};
 
-use crate::{core::config::Config, email::client::EmailClient, routes::health::HealthRouter};
+use crate::{
+    core::config::Config,
+    email::client::EmailClient,
+    routes::{auth::AuthRouter, health::HealthRouter},
+};
 
 #[derive(Debug, Clone)]
 pub struct AppState {
@@ -16,6 +20,7 @@ impl AppRouter {
     pub fn new(state: AppState) -> Router {
         Router::new()
             .nest("/health", HealthRouter::new())
+            .nest("/auth", AuthRouter::new())
             .with_state(state)
     }
 }
