+++ common/src/models/generic.rs
@@ -1,0 +1,6 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Serialize, Deserialize)]
+pub struct MessageResponse {
+    pub message: String,
+}
