+++ common/src/models/mod.rs
@@ -1,6 +1,7 @@
 pub mod appearance;
 pub mod company;
 pub mod error;
+pub mod generic;
 pub mod health;
 pub mod job;
 pub mod payment;
